Move controllers back to controllers/ due to gitops-operator
